refactor: 파일 구조 변경에 따른 파일 위치 변경
- 파일 구조 변경에 따른 파일 위치 에러 해결.
