+++ src/components/myTravel/itinerary/addModal/StepPlaceSelection.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import ScheduleCard from "./ScheduleCard";
+import ScheduleCard from "../ScheduleCard";
 
 const StepPlaceSelection = ({
   locations,
