+++ src/components/myTravel/itinerary/addModal/StepScheduleSetup.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import styled from "styled-components";
-import ScheduleCard from "./ScheduleCard";
+import ScheduleCard from "../ScheduleCard";
 import { formatKoreaDate } from "../../../../utils/changeDateFormUtils";
 
 // select data
